working graphics with cubes
