+++ include/Display.h
@@ -3,9 +3,19 @@
 
 #include <math.h>
 #include <iostream>
+#include <vector>
+#include "RandomNumberGenerator.h"
+
+struct Coord{
+    float x, y, z;
+};
 
 class Display{
 private:
+    inline static int N = 1;
+    inline static std::vector<Coord> particles;
+    inline static std::vector<Coord> velocities;
+
     inline static float pos = 0.0f;
     inline static float vel = 0.05f;
 
@@ -20,6 +30,7 @@ private:
     inline static float radius = 5.0f;
 
 public:
+    Display(int N);
 
     static void reshape(int x, int y);
